Don't let hermes pollute the page's jQuery cc:@stecb @lleirborras
To reduce the amount of assets to be loaded and parsed, the Hermes
embed code checks whether we already have jQuery in window and if
it is recent (>= 1.9) enough. If it is, then it is reused, else,
a fresh 1.11.2 jQuery is loaded off Google's CDN.

However, as Hermes uses bootstrap components that alter the jQuery.fn
object, this would hijack the page's jQuery.fn object, leading to
unexpected behaviour in hermes-embedding applications.

As such, this change introduces a jQuery duplication mechanism that
will make Hermes work on a distinct object than the window one, and
this fixes many sources of headaches in applications that already
use Bootstrap components :-).